requirements.txt
langgraph 
langsmith 
langchain_openai
langchain_community
fastapi
chromadb
python-docx
pandas
comtypes
pdflatex